aitestbed/requirements.txt
# 6G AI Traffic Testbed Dependencies

# Network Emulation (local package)
-e ../netemu

# LLM Clients
openai>=1.0.0
google-generativeai>=0.3.0
websockets>=12.0  # For OpenAI Realtime API

# Data Processing
pandas>=2.0.0
pyarrow>=14.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.28.0

# Optional: Image processing for multimodal
Pillow>=10.0.0

# Optional: Token counting
tiktoken>=0.5.0

# L7 Capture (HTTP/HTTPS interception)
mitmproxy>=10.0.0

# L3/L4 Capture (pcap analysis)
dpkt>=1.9.8

# Optional: WebRTC transport for realtime scenarios
aiortc>=1.5.0

# Optional: Computer use executor
playwright>=1.41.0

# MCP (Model Context Protocol) - Note: MCP servers are npm packages
# Install MCP servers via npm: npm install -g @modelcontextprotocol/server-brave-search
# The Python MCP client uses subprocess to communicate with npm-based servers

# Async support (included in Python 3.7+, but useful for type hints)
# asyncio is part of stdlib
